requirements.txt
# tensorflow
keras
streamlit
efficientnet
matplotlib
seaborn
nltk
google-generativeai
pillow
httpx
python-dotenv